src/requirements.txt
pandas
python-binance
ta
python-dotenv